- paging: use int instead of long - changed wizard controls and timer (removed event) - other fixes and improvements

+++ kuharija/src/main/java/si/merljak/magistrska/client/handler/PagingHandler.java
@@ -12,5 +12,5 @@ public interface PagingHandler {
 	 * 
 	 * @param page new page
 	 */
-	void changePage(long page);
+	void changePage(int page);
 }
